+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - Add "Cannery" to page titles
 - Don't show true/false column for ammo types if all values are false
 - Fix ammo type firing type display
-
+- Show original count, current value, and percentage remaining for ammo groups
+- Show shot history for an ammo group
 
 # 0.1.0
 - Initial release!